File management
# New type : DirectusFile
This file can be used to easily download files from their id.
Just create the object with the id of the file and use the methods to generate the download url with optionnal parameters.
The base url is a static property set automatically base the DirectusApiManager base object when created since it receives this base url for the API.

# Features
- Upload new files
- Update existing File
- Import new File from URL
- automated tests