Breaking : Manage public authorization on new user
If we can allow a unlogged user to create an account for our app, we can not allow a unlogged to get any user in our database.

So I think it will be better to return a boolean on the account creation. If it is true, then we can logged the user in and provide his account.