Bug Fix
To use particular collection like directus_users or directus_file, the websocket was working or the previous api call but not both as the same time.

For example, to call a file via the api we have to use https://domain/files. With the web socket, we must call https://domain/websocket/directus_files

To solve this, I have added a webSocketEndPoint optional string. If it provided, it will be used. The endPointName will be used otherwise.