Features : improve updateItem
to specify the fields that can be updated if there is difference with the fields that can be seen.

For example, if on a collection the user can see a field but is not allowed to update it, then updateItem will return a 403 error because there is difference between the two cases.